Add optional 'name' parameter to expression constructors
Introduce an optional `name` parameter to the constructors of
`ConnectedExpression` and `FilterExpression` classes. This allows
assignment of a name to filter expressions. Updated constructors
in `ConnectedExpression.cs` and `FilterExpression.cs` to accept
`string? name` and set the `Name` property if not null. Changed
`Name` property in `FilterExpression` from `init` to `set` for
modification post-initialization. Ensured backward compatibility
by providing a default value of `null` for the new parameter.